Additional cleanup to chess
git-svn-id: file:///home/runner/work/tom7misc-sync-action/tom7misc-sync-action/upstream/trunk@6215 c7e11af0-4c7a-11f1-9edd-117cb6f09d55

+++ cc-lib/ansi.cc
@@ -445,7 +445,7 @@ ANSI::Decompose(const std::string &text_with_codes,
 
             if (VERBOSE) {
               printf("Pidx: %d\n", pidx);
-              for (int i = 0; i < pleft.size(); i++) {
+              for (int i = 0; i < (int)pleft.size(); i++) {
                 printf("Pleft[%d] = %d\n", i, pleft[i]);
               }
             }
